libs/hfdm/pyproject.toml
[project]
name = "hfdm"
version = "0.1.0"
description = "A module to help me deploying a HugggingFace docker space from and ad hoc folder"
readme = "README.md"
authors = [
    { name = "Christophe GUILLOT", email = "cguillot.abo@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=0.29.3",
    "omegaconf>=2.3.0",
    "pyaml>=25.1.0",
]
classifiers = ["Private :: Do Not Upload"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hfdm"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
